# swagger.yaml
swagger: "2.0"
info:
  description: |
    Rest services for Recipe Rex web application.
    
  version: "1.0.0"
  title: Recipe Rex API
  termsOfService: http://helloreverb.com/terms/
  contact:
    name: pcottenservices@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: localhost
basePath: /rest
schemes:
  - http
paths:
  /recipes/{recipeId}:
    post:
      summary: Add a new recipe to the database
      description: ""
      operationId: addRecipe
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
        - in: path
          name: recipeId
          description: Recipe id to delete
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: Recipe object that needs to be added to the database
          required: false
          schema:
            $ref: "#/definitions/Recipe"
      responses:
        "405":
          description: Invalid input
    delete:
      summary: Deletes a recipe
      description: ""
      operationId: deleteRecipe
      produces:
        - application/json
        - application/xml
      parameters:
        - in: path
          name: recipeId
          description: Recipe id to delete
          required: true
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid recipe id

    put:
      summary: Update an existing recipe
      description: ""
      operationId: updateRecipe
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
        - in: path
          name: recipeId
          description: Recipe id to update
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: Recipe object that needs to be updated in the database
          required: false
          schema:
            $ref: "#/definitions/Recipe"
      responses:
        "405":
          description: Validation exception
        "404":
          description: Recipe not found
        "400":
          description: Invalid ID supplied
...
  /users/{username}/cookbooks/{cookbookId}/recipes/{recipeId}:
    post:
      summary: Adds a recipe to the specified cookbook to the specified user's collection
      description: ""
      operationId: addRecipeToCookbook
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
        - in: path
          name: username
          description: Username associated with collection of cookbooks
          required: true
          type: string
        - in: path
          name: cookbookId
          description: Id of the cookbook to be created
          required: true
          type: string
        - in: path
          name: recipeId
          description: Id of the recipe to be added to the cookbook
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: successful operation
    put:
      summary: Updates a recipe within the specified cookbook in the specified user's collection
      description: ""
      operationId: updateRecipeInCookbook
      consumes:
        - application/json
        - application/xml
      parameters:
        - in: path
          name: username
          description: Username associated with collection of cookbooks
          required: true
          type: string
        - in: path
          name: cookbookId
          description: Id of the cookbook to be created
          required: true
          type: string
        - in: path
          name: recipeId
          description: Id of the recipe to be added to the cookbook
          required: true
          type: integer
          format: int64 
      responses:
        "200":
          description: successful operation 
    delete:
      summary: Updates a recipe within the specified cookbook in the specified user's collection
      description: ""
      operationId: deleteRecipeInCookbook
      consumes:
        - application/json
        - application/xml
      parameters:
        - in: path
          name: username
          description: Username associated with collection of cookbooks
          required: true
          type: string
        - in: path
          name: cookbookId
          description: Id of the cookbook to be created
          required: true
          type: string
        - in: path
          name: recipeId
          description: Id of the recipe to be added to the cookbook
          required: true
          type: integer
          format: int64 
      responses:
        "200":
          description: successful operation 
...
  /users/{username}/meals/{mealId}/recipes/{recipeId}:
    post:
      summary: Adds a recipe to the specified meal to the specified user's collection
      description: ""
      operationId: addRecipeToMeal
      consumes:
        - application/json
        - application/xml
      produces:
        - application/json
        - application/xml
      parameters:
        - in: path
          name: username
          description: Username associated with collection of meals
          required: true
          type: string
        - in: path
          name: mealId
          description: Id of the meal the recipe is in
          required: true
          type: string
        - in: path
          name: recipeId
          description: Id of the recipe to be added to the meal
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: successful operation
    put:
      summary: Updates a recipe within the specified meal in the specified user's collection
      description: ""
      operationId: updateRecipelInMeal
      consumes:
        - application/json
        - application/xml
      parameters:
        - in: path
          name: username
          description: Username associated with collection of meals
          required: true
          type: string
        - in: path
          name: mealId
          description: Id of the meal the recipe is in
          required: true
          type: string
        - in: path
          name: recipeId
          description: Id of the recipe to update
          required: true
          type: integer
          format: int64 
      responses:
        "200":
          description: successful operation 
    delete:
      summary: Deletes a recipe from the specified cookbook in the specified user's collection
      description: ""
      operationId: deleteRecipeFromMeal
      consumes:
        - application/json
        - application/xml
      parameters:
        - in: path
          name: username
          description: Username associated with collection of meals
          required: true
          type: string
        - in: path
          name: mealId
          description: Id of the meal the recipe is in
          required: true
          type: string
        - in: path
          name: recipeId
          description: Id of the recipe to be deleted from the meal
          required: true
          type: integer
          format: int64 
      responses:
        "200":
          description: successful operation
...
definitions:
...
  Category:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
  Recipe:
    type: object
    required:
      - name
      - id
    properties:
      id:
        type: integer
        format: int64
      category:
        $ref: "#/definitions/Category"
      name:
        type: string
      attributorId:
        type: integer
        format: int64
      description:
        type: string
      servings:
        type: integer
        format: int64
      ovenTemp:
        type: integer
        format: int64
      instructions:
        type: array
        items: 
          $ref: "#/definitions/Instruction"
      ingredients:
        type: array
        items:
          $ref: "#/definitions/Ingredient"
      status:
        type: string
        description: pet status in the store
...
  Instruction:
    type: object
    properties:
      id:
        type: integer
        format: int64
      instructionOrderIndex:
        type: integer
        format: int64
      instructionText:
        type: string
  Ingredient:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      quantity:
        type: integer
        format: int32